process/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20211102.0=h96cf925_1
aiohttp=3.8.3=py39ha30fb19_0
aiosignal=1.2.0=pyhd8ed1ab_0
anyio=3.6.2=pyhd8ed1ab_0
arrow-cpp=8.0.0=py39had1886b_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
aws-c-common=0.4.57=hb1e8313_1
aws-c-event-stream=0.1.6=h23ab428_5
aws-checksums=0.1.9=hb1e8313_0
aws-sdk-cpp=1.8.185=he271ece_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blosc=1.21.1=h97e831e_3
bokeh=2.4.3=pyhd8ed1ab_3
boost-cpp=1.78.0=h8b082ac_1
branca=0.5.0=pyhd8ed1ab_0
brotli=1.0.9=h5eb16cf_7
brotli-bin=1.0.9=h5eb16cf_7
brotlipy=0.7.0=py39h63b48b0_1004
bzip2=1.0.8=h0d85af4_4
c-ares=1.18.1=h0d85af4_0
ca-certificates=2022.9.24=h033912b_0
cairo=1.16.0=h904041c_1014
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py39hae9ecf2_0
cfitsio=4.1.0=h2c97ad1_0
cftime=1.6.2=py39h7cc1f47_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=py39h6e9494a_0
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_1
cloudpickle=2.2.0=pyhd8ed1ab_0
contourpy=1.0.5=py39h92daf61_0
cramjam=2.5.0=py39h78c2a6c_0
cryptography=38.0.2=py39h7eb6a14_0
curl=7.85.0=h57eb407_0
cycler=0.11.0=pyhd8ed1ab_0
cyrus-sasl=2.1.27=ha724b88_5
cytoolz=0.12.0=py39h701faf5_0
dask=2022.9.0=pyhd8ed1ab_0
dask-core=2022.9.0=pyhd8ed1ab_0
dask-geopandas=0.2.0=pyhd8ed1ab_0
dask-sql=2022.10.0=py39h928cc0e_0
distributed=2022.9.0=pyhd8ed1ab_0
expat=2.4.9=hf0c8a7f_0
fastapi=0.85.1=pyhd8ed1ab_0
fastparquet=0.8.3=py39hf53a84f_0
fiona=1.8.22=py39h6fa385f_0
folium=0.13.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=h5bb23bf_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.37.4=py39ha30fb19_0
freetype=2.12.1=h3f81eb7_0
freexl=1.0.6=hb7f2c08_1
frozenlist=1.3.1=py39h6218fd2_0
fsspec=2022.8.2=pyhd8ed1ab_0
future=0.18.2=py39h6e9494a_5
gdal=3.5.2=py39h6aafa27_1
geopandas=0.11.1=pyhd8ed1ab_0
geopandas-base=0.11.1=pyha770c72_0
geos=3.11.0=hb486fe8_0
geotiff=1.7.1=ha1a2aeb_3
gettext=0.21.1=h8a4c099_0
gflags=2.2.2=hb1e8313_1004
giflib=5.2.1=hbcb3906_2
glog=0.6.0=h8ac2a54_0
greenlet=1.1.3=py39h7a8716b_0
grpc-cpp=1.46.1=h067a048_0
h11=0.14.0=pyhd8ed1ab_0
hdf4=4.2.15=h0623a88_4
hdf5=1.12.2=nompi_hc782337_100
heapdict=1.0.1=py_0
icu=70.1=h96cf925_0
idna=3.4=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=hac89ed1_2
json-c=0.16=h01d06f9_0
kealib=1.4.15=h51f5cc1_1
kiwisolver=1.4.4=py39h7c694c3_0
krb5=1.19.3=hb49756b_0
lcms2=2.12=h577c468_0
lerc=4.0.0=hb486fe8_0
libblas=3.9.0=16_osx64_openblas
libbrotlicommon=1.0.9=h5eb16cf_7
libbrotlidec=1.0.9=h5eb16cf_7
libbrotlienc=1.0.9=h5eb16cf_7
libcblas=3.9.0=16_osx64_openblas
libcurl=7.85.0=h57eb407_0
libcxx=14.0.6=hccf4f1f_0
libdap4=3.20.6=h3e144a0_2
libdeflate=1.14=hb7f2c08_0
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libevent=2.1.10=h815e4d9_4
libffi=3.4.2=h0d85af4_5
libgdal=3.5.2=hcf5fda6_1
libgfortran=5.0.0=10_4_0_h97931a8_25
libgfortran5=11.3.0=h082f757_25
libglib=2.74.0=h3ba3332_0
libiconv=1.17=hac89ed1_0
libkml=1.3.0=haeb80ef_1015
liblapack=3.9.0=16_osx64_openblas
libnetcdf=4.8.1=nompi_hebd45d5_104
libnghttp2=1.47.0=h7cbc4dc_1
libntlm=1.4=h0d85af4_1002
libopenblas=0.3.21=openmp_h429af6e_3
libpng=1.6.38=ha978bb4_0
libpq=14.5=h76c7896_0
libprotobuf=3.20.1=hbc0c0cd_4
librttopo=1.1.0=he07d8f5_11
libspatialindex=1.9.3=he49afe7_4
libspatialite=5.0.1=h611af13_19
libsqlite=3.39.4=ha978bb4_0
libssh2=1.10.0=h7535e13_3
libthrift=0.15.0=h054ceb0_0
libtiff=4.4.0=hdb44e8a_4
libwebp-base=1.2.4=h775f41a_0
libxcb=1.13=h0d85af4_1004
libxml2=2.10.3=hb9e07b5_0
libzip=1.9.2=h3ad4413_1
libzlib=1.2.13=hfd90126_4
llvm-openmp=14.0.4=ha654fa7_0
locket=1.0.0=pyhd8ed1ab_0
lz4=4.0.0=py39h263ca4c_2
lz4-c=1.9.3=he49afe7_1
mapclassify=2.4.3=pyhd8ed1ab_0
markupsafe=2.1.1=py39h63b48b0_1
matplotlib-base=3.6.1=py39hb2f573b_0
msgpack-python=1.0.4=py39h7c694c3_0
multidict=6.0.2=py39h63b48b0_1
munch=2.5.0=py_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h96cf925_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
netcdf4=1.6.1=nompi_py39h0d363ce_100
networkx=2.8.7=pyhd8ed1ab_0
nspr=4.32=hcd9eead_1
nss=3.78=ha8197d3_0
numexpr=2.8.3=py39hf72b562_0
numpy=1.23.4=py39hdfa1d0c_1
openjpeg=2.5.0=h5d0d7b0_1
openssl=1.1.1q=hfd90126_1
orc=1.7.4=h9274d09_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.5.0=py39hecff1ad_0
partd=1.3.0=pyhd8ed1ab_0
pcre=8.45=he49afe7_0
pcre2=10.37=h3f55489_1
pillow=9.2.0=py39h4d560c1_2
pip=22.3=pyhd8ed1ab_0
pixman=0.40.0=hbcb3906_0
poppler=22.04.0=h9d9b264_3
poppler-data=0.4.11=hd8ed1ab_0
postgresql=14.5=h6af9f65_0
proj=9.0.1=h05f0992_1
prompt-toolkit=3.0.31=pyha770c72_0
psutil=5.9.2=py39ha30fb19_0
pthread-stubs=0.4=hc929b4f_1001
pure-sasl=0.6.2=pyhd8ed1ab_0
pyarrow=8.0.0=py39h2202ef3_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.10.2=py39ha30fb19_0
pygeos=0.13=py39h06e4737_0
pygments=2.13.0=pyhd8ed1ab_0
pyhive=0.6.5=pyhd8ed1ab_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyproj=3.4.0=py39h5233054_0
pysocks=1.7.1=pyha2e5f31_6
pytables=3.7.0=py39hc294f97_2
python=3.9.13=h57e37ff_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.21.0=pyhd8ed1ab_0
python-tzdata=2022.5=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2022.4=pyhd8ed1ab_0
pytz-deprecation-shim=0.1.0.post0=py39h6e9494a_2
pyyaml=6.0=py39h63b48b0_4
re2=2022.04.01=h96cf925_0
readline=8.1.2=h3899abd_0
requests=2.28.1=pyhd8ed1ab_1
rtree=1.0.1=py39h7d0d40a_0
sasl=0.3.1=py39hfd1d529_2
scikit-learn=1.1.2=py39h8031a56_0
scipy=1.9.1=py39h9488793_0
setuptools=65.5.0=pyhd8ed1ab_0
shapely=1.8.5=py39hf33cec1_1
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=h6e38e02_1
sniffio=1.3.0=pyhd8ed1ab_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
sqlalchemy=1.4.42=py39ha30fb19_1
sqlite=3.39.4=h9ae0607_0
starlette=0.20.4=pyhd8ed1ab_1
tabulate=0.9.0=pyhd8ed1ab_1
tblib=1.7.0=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
thrift=0.16.0=py39hfd1d529_1
thrift_sasl=0.4.3=pyhd8ed1ab_2
tiledb=2.11.3=h3b7b576_0
tk=8.6.12=h5dbffcc_0
toolz=0.12.0=pyhd8ed1ab_0
tornado=6.1=py39h63b48b0_3
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzcode=2022e=hb7f2c08_0
tzdata=2022e=h191b570_0
tzlocal=4.2=py39h6e9494a_1
unicodedata2=14.0.0=py39h63b48b0_1
urllib3=1.26.11=pyhd8ed1ab_0
utf8proc=2.6.1=h9ed2024_0
uvicorn=0.19.0=py39h6e9494a_0
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.1=pyhd8ed1ab_0
xarray=2022.10.0=pyhd8ed1ab_0
xerces-c=3.2.3=hf5b2a72_5
xorg-libxau=1.0.9=h35c211d_0
xorg-libxdmcp=1.1.3=h35c211d_0
xyzservices=2022.9.0=pyhd8ed1ab_0
xz=5.2.6=h775f41a_0
yaml=0.2.5=h0d85af4_2
yarl=1.7.2=py39h63b48b0_2
zict=2.2.0=pyhd8ed1ab_0
zlib=1.2.13=hfd90126_4
zstd=1.5.2=hfa58983_4
